% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-exp.R
\name{check_exp}
\alias{check_exp}
\title{Diagnostic plots for exponential growth assumption}
\usage{
check_exp(lmm_obj, type = c("time", "fitted", "qq"))
}
\arguments{
\item{lmm_obj}{An lmm object created by lmm().}

\item{type}{Diagnostic plot type. Options are:
"time" for residuals vs time,
"fitted" for residuals vs fitted values,
"qq" for QQ plot of residuals.}
}
\value{
A ggplot object.
}
\description{
Diagnostic plots for exponential growth assumption
}
\examples{
data(melanoma1)
melanoma1$months <- melanoma1$Day / (365/12)
mel1 <- tumr(melanoma1, ID, months, Volume, Treatment)
lmm1 <- lmm(mel1)
check_exp(lmm1)

}
