% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\name{gam}
\alias{gam}
\title{Generalized Additive Mixed Model for Tumor Growth Data}
\usage{
gam(
  tumr_obj = NULL,
  formula = NULL,
  data = NULL,
  id = NULL,
  time = NULL,
  measure = NULL,
  group = NULL,
  n_grid = 20,
  ...
)
}
\arguments{
\item{tumr_obj}{A \code{tumr} object created by \code{\link{tumr}}.}

\item{formula}{GAM fixed-effects formula. Defaults to
\code{log(measure + 1) ~ group + s(time, by = group)}.}

\item{data}{A data frame of tumor growth measurements.}

\item{id}{Column name of subject IDs.}

\item{time}{Column name of repeated time measurements.}

\item{measure}{Column name of repeated tumor measurements.}

\item{group}{Column name of treatment group assignments.}

\item{n_grid}{Number of time grid points for pairwise contrasts. Default \code{20}.}

\item{...}{Further arguments passed to \code{\link[gamm4]{gamm4}}.}
}
\value{
An object of class \code{"gam"} with components \code{relevant_info},
\code{formula}, \code{fit} (the \code{gamm4} object), and \code{contrast_df}
(pairwise group contrasts with SE and 95\\% CI at each time grid point).
}
\description{
Generalized Additive Mixed Model for Tumor Growth Data
}
\examples{
data(melanoma1)
mel1 <- tumr(melanoma1, ID, Day, Volume, Treatment)
gam(mel1)

}
