% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-limit.R
\name{get_limit}
\alias{get_limit}
\title{Calculate the one-half of the minimum nonzero volume}
\usage{
get_limit(x)
}
\arguments{
\item{x}{The tumr object.}
}
\value{
A numeric value as the lower detection limit.
}
\description{
Calculates a lower detection limit based on the smallest nonzero observed
tumor volume. The limit is defined as one-half of the minimum nonzero
\code{Volume} value in the data.
}
\examples{
data(melanoma1)
melanoma1$months <- melanoma1$Day / (365/12)
mel1 <- tumr(melanoma1, ID, months, Volume, Treatment)
get_limit(mel1)

}
