% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-median.R
\name{get_lld}
\alias{get_lld}
\title{Get lower limit of detection}
\usage{
get_lld(x, lld = NULL)
}
\arguments{
\item{x}{Numeric vector of measurements}

\item{lld}{Optional user-supplied lower limit of detection}
}
\value{
A positive numeric lower limit of detection
}
\description{
Get lower limit of detection
}
