% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Melanoma1}
\alias{Melanoma1}
\title{Tumor growth data for a melanoma cancer model (A375)}
\format{
A data frame with 600 rows and 4 variables:
\describe{
  \item{ID}{Mouse identifier to indicate repeated measurements on the same mouse. (factor)}
  \item{Day}{Day of measurment (integer)}
  \item{Volume}{Tumor volume (double)}
  \item{Treatment}{Treatment group (factor). A = control, B = Vemurafenib (BRAF inhibitor), C = Buthionine sulfoximine (BSO), D = Vemurafenib + BSO.}
}
}
\source{
University of Iowa, Holden Cancer Center, experiment SLB19-022.
}
\usage{
Melanoma1
}
\description{
The A375 cell line (metastatic melanoma) was used to create CDX (Cell Line Derived Xenograft) mice.
}
\keyword{datasets}
