% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.R
\name{mixed}
\alias{mixed}
\title{Analysis based on mixed model}
\usage{
mixed(Data)
}
\arguments{
\item{Data}{From gendat}
}
\value{
A p-value
}
\description{
Analysis based on mixed model
}
\examples{
Data <- gendat(5, 2, 6)
mixed(Data)

}
