% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lmm.R
\name{plot.lmm}
\alias{plot.lmm}
\title{Creates Plots of an lmm object}
\usage{
\method{plot}{lmm}(x, type = c("predict", "predict_fold", "contrast"), ...)
}
\arguments{
\item{x}{lmm object}

\item{type}{Character string specifying which plot to produce. One of
\code{"predict"}, \code{"predict_fold"} and \code{"contrast"}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list of ggplot objects.
}
\description{
Creates Plots of an lmm object
}
\examples{
melanoma1$months <- melanoma1$Day / (365/12)
mel1 <- tumr(melanoma1, ID, months, Volume, Treatment)
mel1_lmm <- lmm(mel1)
plot(mel1_lmm, "predict")
plot(mel1_lmm, "predict_fold")
plot(mel1_lmm, "contrast")

}
