% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-quad.R
\name{plot.quad}
\alias{plot.quad}
\title{Plot the result of Exponential quadratic model}
\usage{
\method{plot}{quad}(x, type = c("predict", "contrast"), n_grid = 20, ...)
}
\arguments{
\item{x}{An object of class \code{"quad"}.}

\item{type}{Type of plot to produce. Either \code{"predict"} for fitted
treatment-specific growth curves or \code{"contrast"} for pairwise
treatment contrasts over time. Default is \code{"predict"}.}

\item{n_grid}{Number of time points used to construct the fitted curves.
Default is 20.}

\item{...}{Additional arguments passed to plotting functions.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots fitted treatment-specific quadratic growth curves or pairwise
treatment contrasts over time from a fitted \code{quad} object.
}
\examples{
data(melanoma1)
melanoma1$months <- melanoma1$Day / (365/12)
mel1 <- tumr(melanoma1, ID, months, Volume, Treatment)
quad_obj <- quad(mel1)
plot(quad_obj, type = "predict")
plot(quad_obj, type = "contrast")

}
