% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_median.R
\name{plot_median_para}
\alias{plot_median_para}
\title{Plot of tumor growth over time using parametric method}
\usage{
plot_median_para(
  tumr_obj = NULL,
  data = NULL,
  group = NULL,
  time = NULL,
  measure = NULL,
  id = NULL
)
}
\arguments{
\item{tumr_obj}{takes tumr_obj created by tumr()}

\item{data}{tumor growth data}

\item{group}{Column specifying the treatment group}

\item{time}{Column of repeated time measurements}

\item{measure}{Column of repeated measurements of tumor}

\item{id}{Column of subject ID's}
}
\value{
A ggplot object
}
\description{
Plot of tumor growth over time using parametric method
}
\examples{
data(melanoma2)
mel2 <- tumr(melanoma2, ID, Day, Volume, Treatment)
plot_median_para(mel2) + ggplot2::coord_cartesian(ylim = c(0, 2000))
}
