% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\title{Plot of tumor growth over time}
\usage{
plots(data, group, time, measure, id, stat = median, remove_na = FALSE)
}
\arguments{
\item{data}{tumor growth data}

\item{group}{Column specifying the treatment group for each measurement}

\item{time}{Column of repeated time measurements}

\item{measure}{Column of repeated measurements of tumor}

\item{id}{Column of subject ID's}

\item{stat}{method of data summary for each group}

\item{remove_na}{removes NA values in measure column before plotting}
}
\value{
A plot
}
\description{
Plot of tumor growth over time
}
\examples{
data(breast)
plots(breast, Treatment, Week, Volume, ID, stat = median, remove_na = FALSE)
plots(breast, Treatment, Week, Volume, ID, remove_na = TRUE)
data(melanoma1)
plots(melanoma1, Treatment, Day, Volume, ID)
}
