% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lmm.R
\name{print.lmm}
\alias{print.lmm}
\title{Prints relevant output from lmm}
\usage{
\method{print}{lmm}(x, ...)
}
\arguments{
\item{x}{lmm object}

\item{...}{Other parameters}
}
\value{
relevant output
}
\description{
Prints relevant output from lmm
}
