% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rfeat.R
\name{print.rfeat}
\alias{print.rfeat}
\title{Prints relevant output from rfeat}
\usage{
\method{print}{rfeat}(x, ...)
}
\arguments{
\item{x}{rfeat object}

\item{...}{Other parameters}
}
\value{
relevant output
}
\description{
Prints relevant output from rfeat
}
