% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-median.R
\name{process_data}
\alias{process_data}
\title{Process data to be used in plot(tumr_obj)}
\usage{
process_data(time, measure)
}
\arguments{
\item{time}{vector of time measurements}

\item{measure}{vector of tumor growth measurements}
}
\value{
A list with cleaned measurements and event indicators
}
\description{
Process data to be used in plot(tumr_obj)
}
