% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{prostate}
\alias{prostate}
\title{Tumor growth data for a prostate cancer model}
\format{
A data frame with 336 rows and 6 variables:
\describe{
\item{ID}{Mouse identifier to indicate repeated measurements on the same mouse. (factor)}
\item{Age}{Age of the mouse, in weeks (integer)}
\item{BLI}{Tumor volume, as measured by bioluminescent imaging, in photons (double)}
\item{Genotype}{p53 genotype (factor). WT = wild type, HET = single knockout, DOKO = double knockout.}
}
}
\source{
https://doi.org/10.1371/journal.pone.0232807
}
\usage{
prostate
}
\description{
In this experiment, 49 mice were split into 3 groups. All of the groups have
Pten, a cancer gene, knocked out of their prostates. One group had normal p53,
one group had one copy of the p53 knocked out, and the third group had both
copies of p53 knocked out of their prostates. These mice also had a
bioluminescent reporter so that tumor progression could be monitored by imaging.
}
\keyword{datasets}
