% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfeat.R
\name{rfeat}
\alias{rfeat}
\title{Analysis based on response features}
\usage{
rfeat(Data, linear = TRUE)
}
\arguments{
\item{Data}{From gendat}

\item{linear}{Assume linear growth?}
}
\value{
A p-value
}
\description{
Analysis based on response features
}
