% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmanova.R
\name{rmanova}
\alias{rmanova}
\title{Analysis based on repeated measures ANOVA}
\usage{
rmanova(Data)
}
\arguments{
\item{Data}{From gendat}
}
\value{
A p-value
}
\description{
Analysis based on repeated measures ANOVA
}
\examples{
Data <- gendat(5, 2, 6)
rmanova(Data)

}
