% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-bhm.R
\name{summary.bhm}
\alias{summary.bhm}
\title{Creates a summary of a bhm object}
\usage{
\method{summary}{bhm}(object, ...)
}
\arguments{
\item{object}{bhm object}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list containing posterior summaries on log scale and exp scale
}
\description{
Creates a summary of a bhm object
}
\examples{
\dontrun{
data(melanoma2)
melanoma2$months <- melanoma2$Day / (365/12)
mel2 <- tumr(melanoma2, ID, months, Volume, Treatment)
fit <- bhm(mel2)
summary(fit)
}
}
