% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-lmm.R
\name{summary.lmm}
\alias{summary.lmm}
\title{Creates a summary of an lmm object}
\usage{
\method{summary}{lmm}(object, ...)
}
\arguments{
\item{object}{lmm object}

\item{...}{further arguments passed to or from other methods}
}
\value{
a summary of the lmm object
}
\description{
Creates a summary of an lmm object
}
\examples{
melanoma1$months <- melanoma1$Day / (365/12)
mel1 <- tumr(melanoma1, ID, months, Volume, Treatment)
mel1_lmm <- lmm(mel1)
summary(mel1_lmm)

}
