% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumr_obj.R
\name{tumr}
\alias{tumr}
\title{Create tumr object}
\usage{
tumr(id, time, measure, group)
}
\arguments{
\item{id}{Column of subject ID's}

\item{time}{Column of repeated time measurements}

\item{measure}{Column of repeated measurements of tumor}

\item{group}{Column specifying the treatment group for each measurement}
}
\value{
A tumr object
}
\description{
Create tumr object
}
\examples{
data(breast)
tumr(ID, Week, Volume, Treatment)

}
