% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumr-power.R
\name{tumr_power}
\alias{tumr_power}
\title{Power calculation for tumor growth}
\usage{
tumr_power(n, effect_size, N = 1000, ...)
}
\arguments{
\item{n}{Sample size per group.}

\item{effect_size}{Treatment effect size.}

\item{N}{Number of simulations. Default is 1000.}

\item{...}{Additional arguments for data generation.}
}
\value{
A data frame containing simulation settings and p-values.
}
\description{
Estimates power using simulated tumor growth data and a response-feature
analysis.
}
\examples{
res <- tumr_power(8, 2, N = 10)
mean(res$p < 0.05)

}
