% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumrPower.R
\name{tumrPower}
\alias{tumrPower}
\title{Power calculation for tumor growth}
\usage{
tumrPower(n, effect.size, N = 1000, ...)
}
\arguments{
\item{n}{Sample size}

\item{effect.size}{Vector of effect sizes to consider}

\item{N}{Number of simulations (default 1000)}

\item{...}{Other arguments to gendat}
}
\value{
An array of p-values
}
\description{
Power calculation for tumor growth
}
\examples{
p <- tumrPower(8, 2, 6)

}
