% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhm.R
\name{bhm}
\alias{bhm}
\title{Bayesian hierarchical model for tumor growth}
\usage{
bhm(data, diagnostics = FALSE, return_fit = TRUE, ...)
}
\arguments{
\item{data}{data.frame with columns ID, Day, Volume, Treatment}

\item{diagnostics}{logical; whether to return diagnostic summary}

\item{return_fit}{logical; whether to return the CmdStan fit object}

\item{...}{further arguments passed to CmdStanR $sample()}
}
\value{
A list of posterior summaries (and optionally diagnostics / fit)
}
\description{
Bayesian hierarchical model for tumor growth
}
\examples{
\dontrun{
data(melanoma2)
fit <- bhm(melanoma2)
fit
}
}
