% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double-time.R
\name{dtime}
\alias{dtime}
\title{Tumor Doubling Time Based on Fitted Tumor Growth Model}
\usage{
dtime(x)
}
\arguments{
\item{x}{A fitted model object.}
}
\value{
A list with three components:
\itemize{
\item \code{method}: character string describing the method used.
\item \code{message}: character string indicating that the model should demonstrate an exponential growth pattern.
\item \code{summary}: a data frame summarizing tumor doubling time by treatment group, including mean, median, and 95\\% intervals.
}
}
\description{
Tumor Doubling Time Based on Fitted Tumor Growth Model
}
\examples{
data(melanoma2)
mel2 <- tumr(melanoma2, ID, Day, Volume, Treatment)
fit1 <- bhm(melanoma2)
dtime(fit1)
fit2 <- lmm(mel2)
dtime(fit2)

}
