% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm}
\alias{lmm}
\title{Linear Mixed Model for Tumor Growth Data}
\usage{
lmm(
  tumr_obj = NULL,
  formula = NULL,
  data = NULL,
  id = NULL,
  time = NULL,
  measure = NULL,
  group = NULL
)
}
\arguments{
\item{tumr_obj}{takes tumr_obj created by tumr()}

\item{formula}{linear mixed model formula}

\item{data}{tumor growth data}

\item{id}{Column of subject ID's}

\item{time}{Column of repeated time measurements}

\item{measure}{Column of repeated measurements of tumor}

\item{group}{Column specifying the treatment group for each measurement}
}
\value{
summary of linear mixed model fit
}
\description{
Linear Mixed Model for Tumor Growth Data
}
\examples{
data(melanoma1)
mel1 <- tumr(melanoma1, ID, Day, Volume, Treatment)
lmm(mel1)

lmm(
tumr_obj = meta_data,
formula = "Volume ~ Day + (1 | ID)"
)

data(breast)
lmm(
data = breast,
id = "ID",
group = "Treatment",
time = "Week",
measure = "Volume"
)

}
