% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-quad.R
\name{plot.quad}
\alias{plot.quad}
\title{Plot Quadratic Linear Mixed Model Treatment Contrasts}
\usage{
\method{plot}{quad}(x, ...)
}
\arguments{
\item{x}{An object of class \code{quad}.}

\item{...}{Additional arguments passed to plotting functions.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots pairwise treatment contrasts over time from a fitted \code{quad} object.
}
\examples{
data(melanoma1)
mel1 <- tumr(melanoma1, ID, Day, Volume, Treatment)
quad_obj <- quad(mel1)
plot(quad_obj)

}
