% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rfeat.R
\name{plot.rfeat}
\alias{plot.rfeat}
\title{Plot based on rfeat analysis}
\usage{
\method{plot}{rfeat}(x, ...)
}
\arguments{
\item{x}{An rfeat object}

\item{...}{Other parameters}
}
\value{
A plot
}
\description{
Plot based on rfeat analysis
}
\examples{
example <- rfeat(breast, "ID", "Week", "Volume", "Treatment",
transformation = log1p, comparison = "t.test")
plot(example)
}
