% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-median.R
\name{plot.tumr}
\alias{plot.tumr}
\title{Plot a tumr object}
\usage{
\method{plot}{tumr}(x, ...)
}
\arguments{
\item{x}{A tumr object created by \code{tumr()}.}

\item{...}{Additional arguments passed to \code{plot_median()}.}
}
\value{
A ggplot object.
}
\description{
Plot a tumr object
}
