% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gam.R
\name{plot.tumr_gam}
\alias{plot.tumr_gam}
\title{Plot of GAM Fit}
\usage{
\method{plot}{tumr_gam}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"tumr_gam"} returned by \code{\link{gam}}.}

\item{...}{Currently ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots estimated pairwise treatment contrasts over time with 95\\% confidence
intervals, faceted by group pair.
}
