% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gam.R
\name{plot.tumr_gam}
\alias{plot.tumr_gam}
\title{Plot GAM Fit for Tumor Growth Data}
\usage{
\method{plot}{tumr_gam}(x, type = c("predict", "contrast"), n_grid = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"tumr_gam"}.}

\item{type}{Either \code{"predict"} (fitted curves per group) or
\code{"contrast"} (pairwise differences over time).}

\item{n_grid}{Number of time grid points. Defaults to \code{x$n_grid}.}

\item{...}{Currently ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot GAM Fit for Tumor Growth Data
}
