% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-median.R
\name{plot_median}
\alias{plot_median}
\title{Plot tumor volume or fold change over time using median curves}
\usage{
plot_median(
  tumr_obj = NULL,
  data = NULL,
  group = NULL,
  time = NULL,
  measure = NULL,
  id = NULL,
  par = TRUE,
  fold = FALSE,
  lld = NULL
)
}
\arguments{
\item{tumr_obj}{Optional tumr object created by tumr()}

\item{data}{Tumor growth data}

\item{group}{Column specifying treatment group}

\item{time}{Column of repeated time measurements}

\item{measure}{Column of repeated tumor measurements}

\item{id}{Column of subject IDs}

\item{par}{Logical. If TRUE, use parametric median estimation.
If FALSE, use nonparametric Kaplan-Meier estimation.}

\item{fold}{Logical. If TRUE, plot fold change instead of raw volume.}

\item{lld}{Optional lower limit of detection used to replace zero values
when fold = TRUE. If NULL, the minimum nonzero value in the full dataset
is used.}
}
\value{
A ggplot object
}
\description{
Plot tumor volume or fold change over time using median curves
}
\examples{
data(melanoma2)
melanoma2$months <- melanoma2$Day / (365/12)
mel2 <- tumr(melanoma2, ID, months, Volume, Treatment)
plot_median(mel2)
plot_median(mel2, par = FALSE)
plot_median(mel2, fold = TRUE)
plot_median(mel2, par = FALSE, fold = TRUE)

}
