% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_median.R
\name{plot_median}
\alias{plot_median}
\title{Plot of tumor growth over time using the median}
\usage{
plot_median(
  tumr_obj = NULL,
  data = NULL,
  group = NULL,
  time = NULL,
  measure = NULL,
  id = NULL,
  par = TRUE
)
}
\arguments{
\item{tumr_obj}{takes tumr_obj created by tumr()}

\item{data}{tumor growth data}

\item{group}{Column specifying the treatment group for each measurement}

\item{time}{Column of repeated time measurements}

\item{measure}{Column of repeated measurements of tumor}

\item{id}{Column of subject ID's}

\item{par}{Logical. If \code{TRUE}, uses the parametric method. If \code{FALSE},
uses the nonparametric method based on Kaplan-Meier
estimation. Default is \code{TRUE}.}
}
\value{
A ggplot object
}
\description{
Plot of tumor growth over time using the median
}
\examples{
data(melanoma2)
mel2 <- tumr(melanoma2, ID, Day, Volume, Treatment)
plot_median(mel2)
plot_median(mel2, par = FALSE)

}
