% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_median}
\alias{plot_median}
\title{Plot of tumor growth over time using the median}
\usage{
plot_median(data, group, time, measure, id, type = "spaghetti")
}
\arguments{
\item{data}{tumor growth data}

\item{group}{Column specifying the treatment group for each measurement}

\item{time}{Column of repeated time measurements}

\item{measure}{Column of repeated measurements of tumor}

\item{id}{Column of subject ID's}

\item{type}{type of plot produced}
}
\value{
A plot
}
\description{
Plot of tumor growth over time using the median
}
\examples{
data(breast)
plot_median(
data = breast,
group = "Treatment",
time = "Week",
measure = "Volume",
id = "ID"
)
data(melanoma1)
plot_median(
data = melanoma1,
group = "Treatment",
time = "Day",
measure = "Volume",
id = "ID",
type = "conf_int"
)
data(melanoma2)
plot_median(
data = melanoma2,
group = "Treatment",
time = "Day",
measure = "Volume",
id = "ID"
)
data(prostate)
plot_median(
data = prostate,
group = "Genotype",
time = "Age",
measure = "BLI",
id = "ID"
)
}
