% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{process_data}
\alias{process_data}
\title{Plot of tumor growth over time}
\usage{
process_data(time, measure)
}
\arguments{
\item{time}{vector of time measurements}

\item{measure}{vector of tumor growth measurements}
}
\value{
Processed Data
}
\description{
Plot of tumor growth over time
}
