% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_median.R
\name{process_data}
\alias{process_data}
\title{Process data to be used in plot_median}
\usage{
process_data(time, measure)
}
\arguments{
\item{time}{vector of time measurements}

\item{measure}{vector of tumor growth measurements}
}
\value{
Processed Data
}
\description{
Process data to be used in plot_median
}
