% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{process_data}
\alias{process_data}
\title{Plot of tumor growth over time}
\usage{
process_data(x)
}
\arguments{
\item{x}{tumor growth data}
}
\value{
Processed Data
}
\description{
Plot of tumor growth over time
}
