% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad.R
\name{quad}
\alias{quad}
\title{Exponential quadratic Model for Tumor Growth Data}
\usage{
quad(tumr_obj = NULL, n_grid = 20, ...)
}
\arguments{
\item{tumr_obj}{A \code{tumr} object created by \code{\link[=tumr]{tumr()}}.}

\item{n_grid}{Number of time points used to evaluate treatment contrasts.
Default is 20.}

\item{...}{Further arguments passed to \code{\link[lme4:lmer]{lme4::lmer()}}.}
}
\value{
An object of class \code{quad} containing the fitted model,
estimated marginal means, and pairwise treatment contrasts.
}
\description{
Fits an Exponential quadratic model for tumor growth data and computes
pairwise treatment contrasts over a grid of time points using estimated
marginal means.
}
\examples{
data(melanoma1)
melanoma1$months <- melanoma1$Day / (365/12)
mel1 <- tumr(melanoma1, ID, months, Volume, Treatment)
quad_obj <- quad(mel1)

}
