% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad.R
\name{quad}
\alias{quad}
\title{Quadratic Linear Mixed Model for Tumor Growth Data}
\usage{
quad(
  tumr_obj = NULL,
  data = NULL,
  id = NULL,
  time = NULL,
  measure = NULL,
  group = NULL,
  n_grid = 20,
  ...
)
}
\arguments{
\item{tumr_obj}{A \code{tumr} object created by \code{\link[=tumr]{tumr()}}.}

\item{data}{Tumor growth data. Only used if \code{tumr_obj} is not supplied.}

\item{id}{Column name for subject IDs.}

\item{time}{Column name for repeated time measurements.}

\item{measure}{Column name for tumor measurements.}

\item{group}{Column name specifying the treatment group.}

\item{n_grid}{Number of time points used to evaluate treatment contrasts.
Default is 20.}

\item{...}{Further arguments passed to \code{\link[lme4:lmer]{lme4::lmer()}}.}
}
\value{
An object of class \code{quad}, which is a list containing:
\describe{
\item{data}{The processed tumor growth data.}
\item{fit}{The fitted quadratic linear mixed model.}
\item{emm}{Estimated marginal means by treatment at each time point.}
\item{contrast_obj}{Pairwise treatment contrasts.}
\item{contrast_df}{A data frame containing contrast estimates and confidence intervals.}
}
}
\description{
Fits a quadratic linear mixed model for tumor growth data and computes
pairwise treatment contrasts over a grid of time points using estimated
marginal means.
}
\examples{
data(melanoma1)
mel1 <- tumr(melanoma1, ID, Day, Volume, Treatment)
quad_obj <- quad(mel1)

}
