% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfeat.r
\name{rfeat}
\alias{rfeat}
\title{Analysis based on response features}
\usage{
rfeat(data, id, time, measure, group, transformation = NULL, comparison)
}
\arguments{
\item{data}{From gendat}

\item{id}{Column of subject ID's}

\item{time}{Column of repeated time measurements}

\item{measure}{Column of repeated measurements of tumor}

\item{group}{Column specifying the treatment group for each measurement}

\item{transformation}{transformation of measurement}

\item{comparison}{Takes "t.test", "anova", "tukey", or "both"}
}
\value{
A p-value
}
\description{
Volume ~ Week * Trt, separate function that picks apart this
}
\details{
exponential growth, separate packages
}
\examples{
data(breast)
rfeat(breast, "ID", "Week", "Volume", "Treatment", transformation = log1p,
comparison = "t.test")
data(melanoma1)
rfeat(melanoma1, "ID", "Day", "Volume", "Treatment", comparison = "anova")
rfeat(melanoma1, "ID", "Day", "Volume", "Treatment", comparison = "both")
}
