% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfeat.r
\name{rfeat}
\alias{rfeat}
\title{Analysis based on response features}
\usage{
rfeat(
  tumr_obj = NULL,
  data = NULL,
  id = NULL,
  time = NULL,
  measure = NULL,
  group = NULL,
  transformation = NULL,
  comparison = c("t.test", "anova", "tukey", "both")
)
}
\arguments{
\item{tumr_obj}{takes tumr_obj created by tumr()}

\item{data}{From gendat}

\item{id}{Column of subject ID's}

\item{time}{Column of repeated time measurements}

\item{measure}{Column of repeated measurements of tumor}

\item{group}{Column specifying the treatment group for each measurement}

\item{transformation}{transformation of measurement}

\item{comparison}{Takes "t.test", "anova", "tukey", or "both"}
}
\value{
A p-value
}
\description{
Volume ~ Week * Trt, separate function that picks apart this
}
\details{
exponential growth, separate packages
}
\examples{
data(breast)
rfeat(
data = breast,
id = "ID",
time = "Week",
measure = "Volume",
group = "Treatment",
transformation = log1p,
comparison = "t.test")
data(melanoma1)
rfeat(
data = melanoma1,
id = "ID",
time = "Day",
measure = "Volume",
group = "Treatment",
comparison = "anova")

}
