% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumr-power.R
\name{rfeat_pwr}
\alias{rfeat_pwr}
\title{Analysis based on response features}
\usage{
rfeat_pwr(Data, linear = TRUE)
}
\arguments{
\item{Data}{From gendat}

\item{linear}{Assume linear growth?}
}
\value{
A p-value
}
\description{
Analysis based on response features
}
\examples{
dat <- gendat(5, 2, 6)
rfeat_pwr(dat)

}
