% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lmm.R
\name{summary.lmm}
\alias{summary.lmm}
\title{Creates a summary of an lmm object}
\usage{
\method{summary}{lmm}(x)
}
\arguments{
\item{x}{lmm object}
}
\value{
a summary of the lmm object
}
\description{
Creates a summary of an lmm object
}
\examples{
mel1 <- tumr(melanoma1, ID, Day, Volume, Treatment)
mel1_lmm <- lmm(mel1)
summary(mel1_lmm)

}
