% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumr-package.R
\docType{package}
\name{tumr-package}
\alias{tumr-package}
\title{tumr: Analysis and Power Calculations for Tumor Growth Rates
Analysis and power calculations for tumor growth rates.}
\description{
This package contains functions for tumor growth rate experiments: plotting and analyzing existing data, simulating new data, and power/sample size calculations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pbreheny.github.io/tumr/}
  \item \url{https://github.com/pbreheny/tumr}
}

}
\author{
\strong{Maintainer}: Patrick Breheny \email{patrick-breheny@uiowa.edu} (\href{https://orcid.org/0000-0002-0650-1119}{ORCID})

Authors:
\itemize{
  \item Tian Zhao \email{tian-zhao@uiowa.edu} (\href{https://orcid.org/0009-0006-9527-286X}{ORCID})
  \item Sydney Plowman
}

}
