% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumr.r
\docType{package}
\name{tumr-package}
\alias{tumr}
\alias{tumr-package}
\title{tumr: Analysis and Power Calculations for Tumor Growth Rates
Analysis and power calculations for tumor growth rates.}
\description{
This package contains functions for tumor growth rate experiments: analyzing existing data, simulating new data, and power/sample size calculations.
}
