% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\name{tumr_gam}
\alias{tumr_gam}
\title{Generalized Additive Model for Tumor Growth Data}
\usage{
tumr_gam(
  tumr_obj = NULL,
  formula = NULL,
  data = NULL,
  id = NULL,
  time = NULL,
  measure = NULL,
  group = NULL,
  n_grid = 20,
  ...
)
}
\arguments{
\item{tumr_obj}{A \code{tumr} object created by \code{\link{tumr}}.}

\item{formula}{GAM formula. Defaults to
\code{log(measure + 1) ~ group + s(time, by = group) +
  s(id, bs = "re") + s(id, time, bs = "re")}.}

\item{data}{A data frame of tumor growth measurements.}

\item{id}{Column name of subject IDs.}

\item{time}{Column name of repeated time measurements.}

\item{measure}{Column name of repeated tumor measurements.}

\item{group}{Column name of treatment group assignments.}

\item{n_grid}{Number of time grid points for prediction. Default \code{20}.}

\item{...}{Further arguments passed to \code{\link[mgcv]{gam}}.}
}
\value{
An object of class \code{"tumr_gam"} with components:
\describe{
\item{\code{fit}}{The \code{mgcv::gam} object.}
\item{\code{data}}{Processed data with internal column names.}
\item{\code{formula}}{The formula used.}
\item{\code{relevant_info}}{Named list of column names and reference group.}
\item{\code{n_grid}}{Number of prediction grid points.}
}
}
\description{
Generalized Additive Model for Tumor Growth Data
}
\examples{
data(melanoma1)
melanoma1$months <- melanoma1$Day / (365/12)
mel1 <- tumr(melanoma1, ID, months, Volume, Treatment)
tumr_gam(mel1)

}
