% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumr-power.R
\name{tumr_power}
\alias{tumr_power}
\title{Power calculation for tumor growth}
\usage{
tumr_power(n, effect_size, N = 1000, ...)
}
\arguments{
\item{n}{Sample size per group (integer or vector)}

\item{effect_size}{As in \code{gendat()} (numeric or vector)}

\item{N}{Number of simulations (default: 1000)}

\item{...}{Other arguments to \code{gendat()}}
}
\value{
An array of p-values
}
\description{
For speed, uses a response feature analysis \code{rfeat()}; this will likely underestimate
power somewhat.
}
\examples{
res <- tumr_power(8, 2, 10)
mean(res$p < 0.05)
}
